Mercedes-warranty-agent/requirements.txt
langchain-openai
langchain-community
langgraph
faiss-cpu
pypdf
python-dotenv
streamlit
python-dotenv
pypdf

